HW04

See merge request xpolace5/pb071!5

+++ hw04/src/currency.h
@@ -1,13 +1,15 @@
 #include "errors.h"
 #include "utils.h"
 
+#include <stdint-gcc.h>
+
 #ifndef CURRENCY_H
 #define CURRENCY_H
 
 struct currency
 {
     char *name;
-    int rating;
+    int64_t rating;
 };
 
 struct currency_table
@@ -20,7 +22,8 @@ struct currency_table
 };
 
 void init_currency_table(struct currency_table *table);
-void add_currency(struct currency_table *table, const char *name, int rating);
-int convert_currency(struct currency_table *table, int amount, const char *currency);
+void add_currency(struct currency_table *table, const char *name, int64_t rating);
+int64_t convert_currency(struct currency_table *table, int64_t amount, const char *currency);
+
 
 #endif
